+++ Активности/Activity/App.config
@@ -2,14 +2,14 @@
 <configuration>
     <configSections>
         <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-            <section name="Активности.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+            <section name="Activity.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
         </sectionGroup>
     </configSections>
     <startup> 
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/>
     </startup>
     <userSettings>
-        <Активности.Properties.Settings>
+        <Activity.Properties.Settings>
             <setting name="set_nec_guild10" serializeAs="String">
                 <value>0</value>
             </setting>
@@ -79,6 +79,6 @@
             <setting name="last_used_day" serializeAs="String">
                 <value>2000-01-01</value>
             </setting>
-        </Активности.Properties.Settings>
+        </Activity.Properties.Settings>
     </userSettings>
 </configuration>
