+++ Активности/Activity/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="Активности.App"
+﻿<Application x:Class="Activity.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Активности"
+             xmlns:local="clr-namespace:Activity"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
