+++ Активности/Activity/App.xaml.cs
@@ -6,12 +6,17 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace Активности
+namespace Activity
 {
     /// <summary>
     /// Логика взаимодействия для App.xaml
     /// </summary>
     public partial class App : Application
     {
+        public static ResourceDictionary dict = new ResourceDictionary();
+        public App()
+        {
+            dict.Source = new Uri("Properties/DataBase.xaml", UriKind.Relative);
+        }
     }
 }
