+++ Активности/Activity/ControlButton.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Controls;
 
-namespace Активности
+namespace Activity
 {
     class ControlButton : Button
     {
