+++ Активности/Activity/Dictionary1.xaml
@@ -1,10 +1,10 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:Активности">
+                    xmlns:local="clr-namespace:Activity">
     <!--<Style x:Key="{x:Type local:ControlButton}" TargetType="{x:Type local:ControlButton}">
         <Setter Property="HorizontalContentAlignment" Value="Center"/>
         <Setter Property="VerticalContentAlignment" Value="Center"/>
-        <Setter Property="FontFamily" Value="/Активности;component/Fonts/#Diablo"/>
+        <Setter Property="FontFamily" Value="/Activity;component/Fonts/#Diablo"/>
         <Setter Property="FontSize" Value="18"/>
         <Setter Property="Margin" Value="0,0,0,5"/>
     </Style>-->
@@ -60,7 +60,7 @@
         <Setter Property="Foreground" Value="{DynamicResource WindowText}"/>
         <Setter Property="HorizontalContentAlignment" Value="Center"/>
         <Setter Property="VerticalContentAlignment" Value="Bottom"/>
-        <Setter Property="FontFamily" Value="/Активности;component/Fonts/#Diablo"/>
+        <Setter Property="FontFamily" Value="/Activity;component/Fonts/#Diablo"/>
         <Setter Property="FontSize" Value="18"/>
         <Setter Property="FontWeight" Value="Bold"/>
         <Setter Property="BorderThickness" Value="1"/>
