+++ Активности/Activity/PageButton.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="Активности.PageButton"
+﻿<UserControl x:Class="Activity.PageButton"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Активности"
+             xmlns:local="clr-namespace:Activity"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <Grid>
