+++ Активности/Activity/Properties/AssemblyInfo.cs
