+++ Активности/Activity/Properties/DataBase.xaml
@@ -1,0 +1,13 @@
+<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+					xmlns:v="clr-namespace:System;assembly=mscorlib"
+                    xmlns:local="clr-namespace:Activity;assembly=Activity">
+    
+    <x:Array x:Key="Quests" Type="local:Quest">
+        <local:Quest/>
+        <local:Quest/>
+        <local:Quest/>
+    </x:Array>
+
+
+</ResourceDictionary>
