+++ Активности/Activity/Properties/Resources.resx
