+++ Активности/Activity/Quest.xaml
@@ -1,0 +1,53 @@
+﻿<UserControl x:Class="Activity.Quest"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Activity"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="300">
+    <Border x:Name="Гильдия_10_кв" Height="150" VerticalAlignment="Top" Margin="2,4,-243,0" Background="#FF534741" BorderBrush="#FF5F3A17" BorderThickness="2">
+        <Grid>
+            <Border Height="30" VerticalAlignment="Top" BorderBrush="#FF362F2D" BorderThickness="0,0,0,2" Background="#FF594125">
+                <Grid>
+                    <Label Content="[Гильдия] Цепочка из 10 квестов" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="10,2,0,2" Foreground="#FFB5B5B5" FontWeight="Bold" FontFamily="/Активности;component/Fonts/#Diablo" FontSize="14"/>
+                    <StackPanel Orientation="Horizontal" HorizontalAlignment="Right">
+                        <Image x:Name="image" Stretch="UniformToFill" Height="28" Width="28" Source="image/Ежедневные.png"/>
+                        <Label x:Name="guild_10_counter" FontSize="20" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" Background="{x:Null}" FontWeight="Bold" FontFamily="/Активности;component/Fonts/#Diablo"/>
+                    </StackPanel>
+                </Grid>
+            </Border>
+            <Grid Margin="0,30,0,0">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="55*"/>
+                    <ColumnDefinition Width="50*"/>
+                    <ColumnDefinition Width="423*"/>
+                </Grid.ColumnDefinitions>
+                <Grid x:Name="ProgressGrid" FlowDirection="RightToLeft" Width="28" HorizontalAlignment="Right" Grid.Column="2">
+                    <ProgressBar x:Name="guild_10" BorderThickness="0" Background="#FFA6A6A6" Height="116" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" Orientation="Vertical" Maximum="10"/>
+                    <local:CounterButton x:Name="guild_10_plus" VerticalAlignment="Top" Height="28" BorderBrush="{x:Null}" Foreground="{x:Null}" HorizontalContentAlignment="Center" Click="guild_10_plus_Click">
+                        <Border Width="26" Height="26" HorizontalAlignment="Center" VerticalAlignment="Center">
+                            <Border.Background>
+                                <ImageBrush ImageSource="image/plus.png" Stretch="UniformToFill"/>
+                            </Border.Background>
+                        </Border>
+                    </local:CounterButton>
+                    <local:CounterButton x:Name="guild_10_minus" VerticalAlignment="Bottom" BorderBrush="{x:Null}" Foreground="{x:Null}" Height="28" Click="guild_10_minus_Click">
+                        <Border Width="26" Height="26" HorizontalAlignment="Center" VerticalAlignment="Center">
+                            <Border.Background>
+                                <ImageBrush ImageSource="image/minus.png" Stretch="UniformToFill"/>
+                            </Border.Background>
+                        </Border>
+                    </local:CounterButton>
+                </Grid>
+                <Label Content="НПЦ:" HorizontalAlignment="Left" VerticalAlignment="Top" Foreground="#FFB5B5B5" FontFamily="/Активности;component/Fonts/#Diablo" FontWeight="Bold" Margin="5,2,0,0"/>
+                <Label Content="Расположение:" HorizontalAlignment="Left" VerticalAlignment="Top" Foreground="#FFB5B5B5" FontFamily="/Активности;component/Fonts/#Diablo" FontWeight="Bold" Margin="5,31,0,0" Grid.ColumnSpan="2" Width="102"/>
+                <Label Content="Требования:" HorizontalAlignment="Left" VerticalAlignment="Top" Foreground="#FFB5B5B5" FontFamily="/Активности;component/Fonts/#Diablo" FontWeight="Bold" Margin="5,60,0,0" Grid.ColumnSpan="2" Width="92"/>
+                <Label Content="Трактирщик" HorizontalAlignment="Left" VerticalAlignment="Top" Foreground="#FFB5B5B5" FontFamily="/Активности;component/Fonts/#Diablo" FontWeight="Bold" Margin="0,2,0,0" Grid.Column="2"/>
+                <Label Content="190, -105, 65 (Ги База)" HorizontalAlignment="Left" VerticalAlignment="Top" Foreground="#FFB5B5B5" FontFamily="/Активности;component/Fonts/#Diablo" FontWeight="Bold" Margin="0,31,0,0" Grid.Column="2"/>
+                <Label Content="30 лвл" HorizontalAlignment="Left" VerticalAlignment="Top" Foreground="#FFB5B5B5" FontFamily="/Активности;component/Fonts/#Diablo" FontWeight="Bold" Margin="0,60,0,0" Grid.Column="2"/>
+                <TextBlock HorizontalAlignment="Left" TextWrapping="Wrap" VerticalAlignment="Top" Grid.Column="2" Margin="160,5,0,0" Height="105" Width="237" Foreground="#FFB5B5B5" FontFamily="/Активности;component/Fonts/#Diablo" FontWeight="Bold"><Run Text="D"/><Run Text="ополнительная информ"/><Run Text="ация"/><Run Text=":"/><LineBreak/><Run Text=" "/><Run Text="Этот квет "/><Run Text="да"/><Run Text="е"/><Run Text="т не только прогресс в прокачки гильдии"/><Run Text=", "/><Run Text="но и большое кол-во опыта. За гильд квесты вы будете получать вклад, который сможете тратить на различные бафы и плюшки"/><Run Text="."/></TextBlock>
+            </Grid>
+        </Grid>
+    </Border>
+</UserControl>
